Rename BaseApplication -> Application, Application -> GUIApplication and TaskApplication -> TaskGUIApplication.
