+++ pyface/application.py
@@ -7,118 +7,126 @@
 # is also available online at http://www.enthought.com/licenses/BSD.txt
 # Thanks for using Enthought open source!
 
-""" Define a base application class to create the event loop, and launch
-the creation of application windows.
+"""
+This module defines the :py:class:`BaseApplication` class for Pyface, Tasks
+and similar applications.  Although the primary use cases are for GUI
+applications, the :py:class:`BaseApplication` class does not have any explicit
+dependency on GUI code, and can be used for CLI or server applications.
+
+Usual usage is to subclass :py:class:`BaseApplication` overriding at least the
+:py:method:`BaseApplication._run` method, but usually the
+:py:method:`BaseApplication.start` and :py:method:`BaseApplication.stop`
+methods as well.
+
+However the class can be used as-is by listening to the
+:py:attr:`BaseApplication.application_initialized` event and performing
+appropriate work there::
+
+    def do_work():
+        print("Hello world")
+
+    app = BaseApplication()
+    app.on_trait_change(do_work, 'application_initialized')
+
 """
 
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 import logging
-from traceback import format_exception, format_exception_only
+import os
+import sys
 
 from traits.api import (
-    Bool, Instance, List, ReadOnly, Tuple, Undefined, Vetoable, on_trait_change
+    Callable, Directory, Event, HasStrictTraits, Instance, List,
+    ReadOnly, Str, Vetoable, VetoableEvent
 )
-
-from .base_application import (
-    ApplicationEvent, ApplicationExit, BaseApplication
-)
-from .i_dialog import IDialog
-from .i_image_resource import IImageResource
-from .i_splash_screen import ISplashScreen
-from .i_window import IWindow
 
 logger = logging.getLogger(__name__)
 
 
-class Application(BaseApplication):
-    """ A basic Pyface GUI application.
+class ApplicationException(Exception):
+    """ Exception subclass for Application-centric exceptions """
+    pass
 
-    This handles setting up starting up the GUI, window management, logging,
-    and other common features that we want when creating a GUI application.
 
-    Often you will want to subclass this application, but it can be used as-is
-    by hooking a listener to the `application_initialized` event::
+class ApplicationExit(ApplicationException):
+    """ Exception which indicates application should try to exit.
 
-        from pyface.api import ApplicationWindow, HeadingText
-
-        class MainWindow(ApplicationWindow):
-            def _create_contents(self, parent):
-                self._label = HeadingText(parent, text="Hello World")
-                return self._label.control
-
-        def main():
-            application = Application(id='hello-world', name='Hello World')
-            window = MainWindow()
-            application.on_trait_change(
-                lambda: application.add_window(window),
-                'application_initialized'
-            )
-            application.run()
-
+    If no arguments, then assumed to be a normal exit, otherwise the arguments
+    give information about the problem.
     """
+    pass
 
-    # -------------------------------------------------------------------------
-    # 'Application' interface
-    # -------------------------------------------------------------------------
 
-    # Branding ----------------------------------------------------------------
+class ApplicationEvent(HasStrictTraits):
+    """ An event associated with an application """
 
-    #: The splash screen for the application. No splash screen by default
-    splash_screen = Instance(ISplashScreen)
+    #: The application that the event happened to.
+    application = ReadOnly
 
-    #: The about dialog for the application.
-    about_dialog = Instance(IDialog)
+    #: The type of application event.
+    event_type = ReadOnly
 
-    #: Icon for the application
-    icon = Instance(IImageResource)
 
-    # Window management -------------------------------------------------------
+class Application(HasStrictTraits):
+    """ A base class for applications.
 
-    #: Default window size
-    window_size = Tuple((800, 600))
-
-    #: Currently active Window if any
-    active_window = Instance(IWindow)
-
-    #: List of all open windows in the application
-    windows = List(Instance(IWindow))
-
-    #: The Pyface GUI instance for the application
-    gui = ReadOnly
-
-    # Protected interface -----------------------------------------------------
-
-    #: Flag if the exiting of the application was explicitely requested by user
-    # An 'explicit' exit is when the 'exit' method is called.
-    # An 'implicit' exit is when the user closes the last open window.
-    _explicit_exit = Bool(False)
-
-    # Window lifecycle methods -----------------------------------------------
-
-    def add_window(self, window):
-        """ Add a new window to the windows we are tracking. """
-        if window.size == (-1, -1):
-            window.size = self.window_size
-        if not window.title:
-            window.title = self.name
-        if self.icon:
-            window.icon = self.icon
-
-        # Keep a handle on all windows created so that non-active windows don't
-        # get garbage collected
-        self.windows.append(window)
-
-        window.open()
-
-    def do_about(self):
-        """ Display the about dialog, if it exists. """
-        if self.about_dialog is not None:
-            self.about_dialog.open()
+    This class handles the basic lifecycle of an application and a few
+    fundamental facilities.  It is suitable as a base for any application,
+    not just GUI applications.
+    """
 
     # -------------------------------------------------------------------------
     # 'BaseApplication' interface
     # -------------------------------------------------------------------------
+
+    # Branding ----------------------------------------------------------------
+
+    #: Human-readable application name
+    name = Str('Default Application')
+
+    #: Human-readable company name
+    company = Str
+
+    # Infrastructure ----------------------------------------------------------
+
+    # The application's globally unique identifier.
+    id = Str
+
+    #: Application home directory (for preferences, logging, etc.)
+    home = Directory
+
+    #: User data directory (for user files, projects, etc)
+    user_data = Directory
+
+    # Application lifecycle --------------------------------------------------
+
+    #: Fired when the application is starting. Called immediately before the
+    #: start method is run.
+    starting = Event(Instance(ApplicationEvent))
+
+    #: Upon successful completion of the start method.
+    started = Event(Instance(ApplicationEvent))
+
+    #: Fired after the GUI event loop has been started during the run method.
+    application_initialized = Event(Instance(ApplicationEvent))
+
+    #: Fired when the application is starting. Called immediately before the
+    #: stop method is run.
+    exiting = VetoableEvent
+
+    #: Fired when the application is starting. Called immediately before the
+    #: stop method is run.
+    stopping = Event(Instance(ApplicationEvent))
+
+    #: Upon successful completion of the stop method.
+    stopped = Event(Instance(ApplicationEvent))
+
+    # Private interface ------------------------------------------------------
+
+    _old_excepthook = Callable
+
+    # Application lifecycle methods -------------------------------------------
 
     def start(self):
         """ Start the application, setting up things that are required
@@ -127,115 +135,212 @@ class Application(BaseApplication):
         their start method, and should call the superclass start() method
         before doing any work themselves.
         """
-        from pyface.gui import GUI
-
-        ok = super(Application, self).start()
-        if ok:
-            # create the GUI so that the splash screen comes up first thing
-            if self.gui is Undefined:
-                self.gui = GUI(splash_screen=self.splash_screen)
-
-        return ok
-
-    # -------------------------------------------------------------------------
-    # 'BaseApplication' private interface
-    # -------------------------------------------------------------------------
-
-    def _run(self):
-        """ Actual implementation of running the application: starting the GUI
-        event loop.
-        """
-        # Fire a notification that the app is running.  This is guaranteed to
-        # happen after all initialization has occurred and the event loop has
-        # started.  A listener for this event is a good place to do things like
-        # opening application windows and other activities where you want the
-        # event loop running.
-        self.gui.invoke_later(
-            self._fire_application_event, 'application_initialized'
-        )
-
-        # start the GUI - script blocks here
-        self.gui.start_event_loop()
+        logger.info('---- Application starting ----')
         return True
 
-    # Exception handling ------------------------------------------------------
+    def stop(self):
+        """ Stop the application, cleanly releasing resources if possible. """
+        logger.info('---- Application stopping ----')
+        return True
 
-    def _unhandled_exception(self, type, value, traceback):
-        """ Show an error dialog with exception information, if possible. """
-        from pyface.message_dialog import error
+    def run(self):
+        """ Run the application
 
-        informative = """An unexpected error occurred: {}""".format(
-            format_exception_only(type, value))
-        detail = format_exception(type, value, traceback)
-        error("Unhandled Exception", informative=informative, detail=detail)
-
-    def _excepthook(self, type, value, traceback):
-        """ Handle any exceptions not explicitly caught
-
-        Try to display an appropriate error dialog, if possible.  Fail over to
-        the standard excepthook if that doesn't work.
+        Return
+        ------
+        status : bool
+            Whether or not the application ran normally
         """
-        try:
-            if issubclass(type, ApplicationExit):
-                if value.args:
-                    logger.excpetion("Unhandled ApplicationExit")
+        run = stopped = False
+
+        # Start up the application.
+        self._fire_application_event('starting')
+        started = self.start()
+        if started:
+
+            logger.info('---- Application started ----')
+            self._fire_application_event('started')
+
+            try:
+                run = self._run()
+            except ApplicationExit as exc:
+                if exc.args == ():
+                    logger.info("---- ApplicationExit raised ----")
                 else:
-                    logger.info("Unhandled ApplicationExit")
-                self._exit()
-            else:
-                # try to log the exception, could fail eg. if disk full
-                logger.exception("Unhandled exception")
-                self._unhandled_exception(type, value, traceback)
-        except Exception:
-            # something went wrong, use base class excepthook
-            super(Application, self)._excepthook(type, value, traceback)
+                    logger.exception("---- ApplicationExit raised ----")
+                run = (exc.args == ())
+            finally:
+                # Try to shut the application down.
+                self._fire_application_event('stopping')
+                stopped = self.stop()
+                if stopped:
+                    self._fire_application_event('stopped')
+                    logger.info('---- Application stopped ----')
 
-    # Destruction methods -----------------------------------------------------
+        return started and run and stopped
 
-    def _can_exit(self):
-        """ Check with each window to see if it can be closed
+    def exit(self, force=False):
+        """ Exits the application.
 
-        The fires closing events for each window, and returns False if any
-        listener vetos.
+        This method handles a request to shut down the application by the user,
+        eg. from a menu.  If force is False, the application can potentially
+        veto the close event, leaving the application in the state that it was
+        before the exit method was called.
+
+        Parameters
+        ----------
+        force : bool, optional (default False)
+            If set, windows will receive no closing events and will be
+            destroyed unconditionally. This can be useful for reliably tearing
+            down regression tests, but should be used with caution.
+
+        Raises
+        ------
+        ApplicationExit
+            Some subclasses may trigger the exit by raising ApplicationExit.
         """
-        if not super(Application, self)._can_exit():
-            return False
-
-        for window in reversed(self.windows):
-            window.closing = event = Vetoable()
-            if event.veto:
-                return False
+        logger.info('---- Application exit started ----')
+        if force or self._can_exit():
+            try:
+                self._prepare_exit()
+            except Exception:
+                logger.exception("Error preparing for application exit")
+            finally:
+                logger.info('---- Application exit ----')
+                self._exit()
         else:
-            return True
+            logger.info('---- Application exit vetoed ----')
 
-    def _prepare_exit(self):
-        """ Close each window """
-        # ensure copy of list, as we modify original list while closing
-        for window in list(reversed(self.windows)):
-            window.destroy()
-            window.closed = window
+    # Initialization utilities -----------------------------------------------
 
-    def _exit(self):
-        """ Shut down the event loop """
-        self.gui.stop_event_loop()
+    def initialize_application_home(self):
+        """ Set up the home directory for the application
+
+        This is where logs, preference files and other config files will be
+        stored.
+        """
+        if not os.path.exists(self.home):
+            logger.info('Application home directory does not exist, creating')
+            os.makedirs(self.home)
+
+    def install_excepthook(self):
+        """ Install an exception hook to catch unhandled exceptions
+
+        This permits applications to do things like catch and display an error
+        dialog when something unexpectedly goes wrong.
+        """
+        self._old_excepthook = sys.excepthook
+        sys.excepthook = self._excepthook
+        logger.debug('Exception hook installed')
+
+    # Teardown utilities ------------------------------------------------------
+
+    def reset_excepthook(self):
+        """ Install an exception hook to catch unhandled exceptions
+
+        This permits applications to do things like catch and display an error
+        dialog when something unexpectedly goes wrong.
+        """
+        sys.excepthook = self._old_excepthook
+        self._old_excepthook = None
+        logger.debug('Exception hook reset')
 
     # -------------------------------------------------------------------------
     # Private interface
     # -------------------------------------------------------------------------
 
-    # Initialization utilities ------------------------------------------------
+    # Main method -------------------------------------------------------------
 
-    @on_trait_change('windows:activate')
-    def _on_activate_window(self, window, trait, old, new):
-        """ Listener that tracks currently active window.
+    def _run(self):
+        """ Actual implementation of running the application
+
+        This should be completely overriden by applications which want to
+        actually do something.  Usually this method starts an event loop and
+        blocks, but for command-line applications this could be where the
+        main application logic is called from.
         """
-        self.active_window = window
+        # Fire a notification that the app is running.  If the app has an
+        # event loop (eg. a GUI, Tornado web app, etc.) then this should be
+        # fired _after_ the event loop starts using an appropriate callback
+        # (eg. gui.set_trait_later).
+        self._fire_application_event('application_initialized')
+        return True
 
-    # Destruction utilities ---------------------------------------------------
+    # Utilities ---------------------------------------------------------------
 
-    @on_trait_change('windows:closed')
-    def _on_window_closed(self, window, trait, old, new):
-        """ Listener that ensures window handles are released when closed.
+    def _fire_application_event(self, event_type):
+        event = ApplicationEvent(application=self, event_type=event_type)
+        setattr(self, event_type, event)
+
+    # Exception handling ------------------------------------------------------
+
+    def _excepthook(self, type, value, traceback):
+        """ Handle any exceptions not explicitly caught
+
+        This can be overridden by GUI applications to display an appropriate
+        error dialog, if possible.  Keep in mind that the application can be
+        in an arbitrary state when this is called, so anything provided by the
+        user should fail over to the standard excepthook.
         """
-        if window in self.windows:
-            self.windows.remove(window)
+        try:
+            # try to log the exception, could fail eg. if disk full
+            logger.exception(value)
+        finally:
+            # just use standard excepthook in this app
+            sys.__excepthook__(type, value, traceback)
+            # die, in an error state
+            raise SystemExit("Can't log exception in application excepthook.")
+
+    # Destruction methods -----------------------------------------------------
+
+    def _can_exit(self):
+        """ Is exit vetoed by anything?
+
+        The default behaviour is to fire the :py:attr:`exiting` event and check
+        to see if any listeners veto.  Subclasses may wish to override to
+        perform additional checks.
+
+        Returns
+        -------
+        can_exit : bool
+            Return True if exit is OK, False if something vetoes the exit.
+        """
+        self.exiting = event = Vetoable()
+        return not event.veto
+
+    def _prepare_exit(self):
+        """ Do any application-level state saving and clean-up
+
+        Subclasses should override this method.
+        """
+        pass
+
+    def _exit(self):
+        """ Shut down the application
+
+        This is where application event loops and similar should be shut down.
+        """
+        # invoke a normal exit from the application
+        raise ApplicationExit()
+
+    # Traits defaults ---------------------------------------------------------
+
+    def _id_default(self):
+        """ Use the application's directory as the id """
+        from traits.etsconfig.etsconfig import ETSConfig
+        return ETSConfig._get_application_dirname()
+
+    def _home_default(self):
+        """ Default home comes from ETSConfig. """
+        from traits.etsconfig.etsconfig import ETSConfig
+        return os.path.join(ETSConfig.application_data, self.id)
+
+    def _user_data_default(self):
+        """ Default user_data comes from ETSConfig. """
+        from traits.etsconfig.etsconfig import ETSConfig
+        return ETSConfig.user_data
+
+    def _company_default(self):
+        """ Default company comes from ETSConfig. """
+        from traits.etsconfig.etsconfig import ETSConfig
+        return ETSConfig.company
